Break down helpers.ts
