+++ test/keys/index.ts
@@ -1,6 +1,6 @@
 import path from 'node:path';
 
-import { readJsonFromFile } from '../../src/lib/helpers';
+import { readJsonFromFile } from '../../src/cli';
 
 export const getParsedKey = (filename: string): Record<string, unknown> => {
   const filepath = path.join(__dirname, filename);
