+++ test/lib/test_helpers.ts
@@ -5,7 +5,7 @@ import type {JWTVerifyResult } from "jose";
 import { importJWK, jwtVerify } from "jose";
 
 import type { OAuth2Issuer } from "../../src/lib/oauth2-issuer";
-import { privateToPublicKeyTransformer } from "../../src/lib/helpers";
+import { privateToPublicKeyTransformer } from '../../src/lib/jwk-store.keys';
 
 export const verifyTokenWithKey = async (issuer: OAuth2Issuer, token: string, kid: string): Promise<JWTVerifyResult> => {
   const key = issuer.keys.get(kid);
